Introduce an iterator over all types in the program
The new `drgn_type_iterator` API allows users to iterate over all of the
types contained within a program. This iterator is presently *not* flat
(i.e. if `class A` has a member of type `class B` and `class B` isn't
defined elsewhere, `class B` won't appear as a standalone entry in the
iterator). This behavior may be changed in the future.