Handle rare case where a nested type is defined in a separate type unit
Consider the following code:
```
class A {
    ...
    class B {
	...
    };
};
```

It is rare (but possible!) that the DWARF DIE for `class A` is part of a
Type Unit, and that the DWARF DIE for `class B` is part of a _different_
Type Unit. Previously `drgn` did not handle this correctly, resulting in
it reporting `class B` as an incomplete type. This commit fixes that.